#24 lots of render changes and bug fix
Changed from update subresource to map. Fixed a bug on the Drawable
CBuffer binder. The sort key bitfield was all fucked up, fixed that too.
Main menu buttons are done. Fixed a bug on Window too, I was getting
ride of the clirect position, with would screw the resize anchored..

+++ redesigningv0/namespace text/CreateFontTexture.cpp
@@ -313,7 +313,7 @@ bool text::ReadFontDescFromFile( const char * szFontDescFilename_p, BmpFontDesc 
 	GetFileSizeEx( hFile, &size );
 
 	MorpherUnion data;
-	data.m_data = new BYTE[size.QuadPart];
+	data.m_data = new BYTE[(size_t)size.QuadPart];
 	data.SIZE = (int)size.QuadPart;
 	ReadFile( hFile, (LPVOID)data.m_data, data.SIZE, &dwBytesRead, NULL );
 	CloseHandle( hFile);
